fix:Map 이용해 DTO 선택하도록 수정

+++ together-market/src/main/java/likelion/togethermarket/domain/member/service/MemberService.java
@@ -6,6 +6,7 @@ import likelion.togethermarket.domain.market.entity.WishMarket;
 import likelion.togethermarket.domain.market.repository.MarketRepository;
 import likelion.togethermarket.domain.market.repository.WishMarketRepository;
 import likelion.togethermarket.domain.member.dto.request.CustomerModifyReqDto;
+import likelion.togethermarket.domain.member.dto.request.ModifyReq;
 import likelion.togethermarket.domain.member.dto.request.OwnerModifyReqDto;
 import likelion.togethermarket.domain.member.dto.response.*;
 import likelion.togethermarket.domain.member.entity.BlackList;
@@ -16,6 +17,7 @@ import likelion.togethermarket.domain.member.repository.MemberRepository;
 import likelion.togethermarket.domain.shop.entity.Shop;
 import likelion.togethermarket.domain.shop.repository.ShopRepository;
 import lombok.RequiredArgsConstructor;
+import org.apache.catalina.connector.Response;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -77,50 +79,56 @@ public class MemberService {
         return new ResponseEntity<>(HttpStatusCode.valueOf(500));
     }
 
+
+    //회원정보 수정
     @Transactional
-    public ResponseEntity<?> modifyCustomerInfo(Long memberId, CustomerModifyReqDto modifyRequestDto) {
+    public ResponseEntity<?> modifyInfo(Long memberId, ModifyReq modifyReq){
+
         Member member = memberRepository.findById(memberId).orElseThrow();
 
-        wishMarketRepository.deleteAllByMember(member);  // 선호 가게 수정 (전부 없애고, 다시 다 만들어줌)
-        for (Integer i : modifyRequestDto.getFavourite_markets()) {
-            Market market = marketRepository.findById(Long.valueOf(i)).orElseThrow();
-            WishMarket wish = WishMarket.builder().market(market).member(member).build();
-            wishMarketRepository.save(wish);
+        if (member.getMemberRole() == MemberRole.CUSTOMER) {
+            CustomerModifyReqDto modifyRequestDto = CustomerModifyReqDto.builder().modifyReq(modifyReq).build();
+            wishMarketRepository.deleteAllByMember(member);  // 선호 가게 수정 (전부 없애고, 다시 다 만들어줌)
+            for (Integer i : modifyRequestDto.getFavourite_markets()) {
+                Market market = marketRepository.findById(Long.valueOf(i)).orElseThrow();
+                WishMarket wish = WishMarket.builder().market(market).member(member).build();
+                wishMarketRepository.save(wish);
+            }
+            member.modifyNickname(modifyRequestDto.getNickname());
+            member.modifyIntro(modifyRequestDto.getIntroduction());
+
+            ModifyResDto modifyResDto = ModifyResDto.builder()
+                    .nickname(member.getNickName())
+                    .email(member.getEmail())
+                    .introduction(member.getIntroduction())
+                    .profile(member.getProfile())
+                    .is_owner(false).build();
+
+            return new ResponseEntity<ModifyResDto>(modifyResDto, HttpStatusCode.valueOf(200));
+        } else if (member.getMemberRole() == MemberRole.OWNER) {
+            Shop shop = shopRepository.findByMember(member).orElseThrow();
+
+            OwnerModifyReqDto modifyRequestDto = OwnerModifyReqDto.builder().modifyReq(modifyReq).build();
+            // 사장의 닉네임, 소개 + 가게의 이름, open, close 시간 수정
+            member.modifyNickname(modifyRequestDto.getShop_name());
+            member.modifyIntro(modifyRequestDto.getIntroduction());
+            shop.modifyShopInfo(modifyRequestDto.getShop_name(),
+                    modifyRequestDto.getOpening_time(),
+                    modifyRequestDto.getClosing_time());
+
+            ModifyResDto modifyResDto = ModifyResDto.builder()
+                    .email(member.getEmail())
+                    .nickname(member.getNickName())
+                    .profile(member.getProfile())
+                    .introduction(member.getIntroduction())
+                    .is_owner(true).build();
+
+            return new ResponseEntity<ModifyResDto>(modifyResDto, HttpStatusCode.valueOf(200));
+        } else {
+            return new ResponseEntity<String>("잘못된 시도", HttpStatusCode.valueOf(400));
         }
-        member.modifyNickname(modifyRequestDto.getNickname());
-        member.modifyIntro(modifyRequestDto.getIntroduction());
-
-        ModifyResDto modifyResDto = ModifyResDto.builder()
-                .nickname(member.getNickName())
-                .email(member.getEmail())
-                .introduction(member.getIntroduction())
-                .profile(member.getProfile())
-                .is_owner(false).build();
-
-        return new ResponseEntity<ModifyResDto>(modifyResDto, HttpStatusCode.valueOf(200));
     }
 
-    @Transactional
-    public ResponseEntity<?> modifyOwnerInfo(Long memberId, OwnerModifyReqDto modifyRequestDto) {
-        Member member = memberRepository.findById(memberId).orElseThrow();
-
-        Shop shop = shopRepository.findByMember(member).orElseThrow();
-        // 사장의 닉네임, 소개 + 가게의 이름, open, close 시간 수정
-        member.modifyNickname(modifyRequestDto.getShop_name());
-        member.modifyIntro(modifyRequestDto.getIntroduction());
-        shop.modifyShopInfo(modifyRequestDto.getShop_name(),
-                modifyRequestDto.getOpening_time(),
-                modifyRequestDto.getClosing_time());
-
-        ModifyResDto modifyResDto = ModifyResDto.builder()
-                .email(member.getEmail())
-                .nickname(member.getNickName())
-                .profile(member.getProfile())
-                .introduction(member.getIntroduction())
-                .is_owner(true).build();
-
-        return new ResponseEntity<ModifyResDto>(modifyResDto, HttpStatusCode.valueOf(200));
-    }
 
     @Transactional
     public ResponseEntity<?> blockUser(Long memberId, Long blockMemberId) {
